add jwt logic and tests

+++ part4/bloglist/controllers/blogs.js
@@ -3,13 +3,7 @@ const Blog = require('../models/blog');
 const User = require('../models/user');
 const jwt = require('jsonwebtoken');
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '');
-  }
-  return null;
-};
+const getTokenFrom = require('../utils/tokenExtractor');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user');
@@ -31,22 +25,22 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body;
-  // const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-  // if (!decodedToken.id) {
-  //   return response.status(401).json({ error: 'token invalid' });
-  // }
-  // const user = await User.findById(decodedToken.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes ? body.likes : 0,
-    // user: user._id,
+    user: user._id,
   });
   try {
     const savedBlog = await blog.save();
-    // user.blogs = user.blogs.concat(savedBlog._id)
-    // await user.save()
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
     response.status(201).json(savedBlog);
   } catch (error) {
     next(error);
@@ -68,8 +62,16 @@ blogsRouter.put('/:id', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id)
+  if (user.blogs.toString() === blog.id) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
 });
 
 module.exports = blogsRouter;
